Integrate all stuff with utilisateur

+++ src/main/java/com/berd/dev/dtos/CategorieDepenseDto.java
@@ -8,10 +8,10 @@ import lombok.NoArgsConstructor;
 public class CategorieDepenseDto {
     private Integer idCategorieDepense;
     private String libelle;
-    private Long idUtilisateur;
+    private Integer idUtilisateur;
 
     public CategorieDepenseDto(
-            Integer idCategorieDepense, String libelle, Long idUtilisateur) {
+            Integer idCategorieDepense, String libelle, Integer idUtilisateur) {
         setIdCategorieDepense(idCategorieDepense);
         setIdUtilisateur(idUtilisateur);
         setLibelle(libelle);
